+++ internal/provider/provider.go
@@ -378,6 +378,7 @@ func (p *MerakiProvider) Resources(ctx context.Context) []func() resource.Resour
 		NewSwitchDHCPServerPolicyARPInspectionTrustedServerResource,
 		NewSwitchDSCPToCoSMappingsResource,
 		NewSwitchLinkAggregationResource,
+		NewSwitchLinkAggregationsResource,
 		NewSwitchMTUResource,
 		NewSwitchOrganizationPortsProfileResource,
 		NewSwitchOrganizationPortsProfilesAutomationResource,