+++ internal/provider/resource_meraki_organization_policy_object_groups.go
@@ -21,7 +21,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"strconv"
 	"strings"
 
@@ -236,7 +235,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Update(ctx context.Context, req
 			// If the item is present in state, but not in plan, we need to delete it
 			actions = append(actions, meraki.ActionModel{
 				Operation: "destroy",
-				Resource:  plan.getPath() + "/" + url.QueryEscape(itemState.Id.ValueString()),
+				Resource:  plan.getPath() + "/" + itemState.Id.ValueString(),
 				Body:      "{}",
 			})
 		}
@@ -253,7 +252,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Update(ctx context.Context, req
 				if hasChanges {
 					actions = append(actions, meraki.ActionModel{
 						Operation: "update",
-						Resource:  plan.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+						Resource:  plan.getPath() + "/" + item.Id.ValueString(),
 						Body:      item.toBody(ctx, itemState),
 					})
 				}
@@ -301,7 +300,7 @@ func (r *OrganizationPolicyObjectGroupsResource) Delete(ctx context.Context, req
 	for i, item := range state.Items {
 		actions[i] = meraki.ActionModel{
 			Operation: "destroy",
-			Resource:  state.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+			Resource:  state.getPath() + "/" + item.Id.ValueString(),
 			Body:      "{}",
 		}
 	}