+++ internal/provider/resource_meraki_sm_admin_roles.go
@@ -21,7 +21,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"strconv"
 	"strings"
 
@@ -241,7 +240,7 @@ func (r *SMAdminRolesResource) Update(ctx context.Context, req resource.UpdateRe
 			// If the item is present in state, but not in plan, we need to delete it
 			actions = append(actions, meraki.ActionModel{
 				Operation: "destroy",
-				Resource:  plan.getPath() + "/" + url.QueryEscape(itemState.Id.ValueString()),
+				Resource:  plan.getPath() + "/" + itemState.Id.ValueString(),
 				Body:      "{}",
 			})
 		}
@@ -258,7 +257,7 @@ func (r *SMAdminRolesResource) Update(ctx context.Context, req resource.UpdateRe
 				if hasChanges {
 					actions = append(actions, meraki.ActionModel{
 						Operation: "update",
-						Resource:  plan.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+						Resource:  plan.getPath() + "/" + item.Id.ValueString(),
 						Body:      item.toBody(ctx, itemState),
 					})
 				}
@@ -306,7 +305,7 @@ func (r *SMAdminRolesResource) Delete(ctx context.Context, req resource.DeleteRe
 	for i, item := range state.Items {
 		actions[i] = meraki.ActionModel{
 			Operation: "destroy",
-			Resource:  state.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+			Resource:  state.getPath() + "/" + item.Id.ValueString(),
 			Body:      "{}",
 		}
 	}