+++ internal/provider/resource_meraki_switch_stack_routing_interfaces.go
@@ -21,7 +21,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"strconv"
 	"strings"
 
@@ -288,7 +287,7 @@ func (r *SwitchStackRoutingInterfacesResource) Update(ctx context.Context, req r
 			// If the item is present in state, but not in plan, we need to delete it
 			actions = append(actions, meraki.ActionModel{
 				Operation: "destroy",
-				Resource:  plan.getPath() + "/" + url.QueryEscape(itemState.Id.ValueString()),
+				Resource:  plan.getPath() + "/" + itemState.Id.ValueString(),
 				Body:      "{}",
 			})
 		}
@@ -305,7 +304,7 @@ func (r *SwitchStackRoutingInterfacesResource) Update(ctx context.Context, req r
 				if hasChanges {
 					actions = append(actions, meraki.ActionModel{
 						Operation: "update",
-						Resource:  plan.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+						Resource:  plan.getPath() + "/" + item.Id.ValueString(),
 						Body:      item.toBody(ctx, itemState),
 					})
 				}
@@ -353,7 +352,7 @@ func (r *SwitchStackRoutingInterfacesResource) Delete(ctx context.Context, req r
 	for i, item := range state.Items {
 		actions[i] = meraki.ActionModel{
 			Operation: "destroy",
-			Resource:  state.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+			Resource:  state.getPath() + "/" + item.Id.ValueString(),
 			Body:      "{}",
 		}
 	}