+++ internal/provider/resource_meraki_switch_stack_routing_static_routes.go
@@ -21,7 +21,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"net/url"
 	"strconv"
 	"strings"
 
@@ -251,7 +250,7 @@ func (r *SwitchStackRoutingStaticRoutesResource) Update(ctx context.Context, req
 			// If the item is present in state, but not in plan, we need to delete it
 			actions = append(actions, meraki.ActionModel{
 				Operation: "destroy",
-				Resource:  plan.getPath() + "/" + url.QueryEscape(itemState.Id.ValueString()),
+				Resource:  plan.getPath() + "/" + itemState.Id.ValueString(),
 				Body:      "{}",
 			})
 		}
@@ -268,7 +267,7 @@ func (r *SwitchStackRoutingStaticRoutesResource) Update(ctx context.Context, req
 				if hasChanges {
 					actions = append(actions, meraki.ActionModel{
 						Operation: "update",
-						Resource:  plan.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+						Resource:  plan.getPath() + "/" + item.Id.ValueString(),
 						Body:      item.toBody(ctx, itemState),
 					})
 				}
@@ -316,7 +315,7 @@ func (r *SwitchStackRoutingStaticRoutesResource) Delete(ctx context.Context, req
 	for i, item := range state.Items {
 		actions[i] = meraki.ActionModel{
 			Operation: "destroy",
-			Resource:  state.getPath() + "/" + url.QueryEscape(item.Id.ValueString()),
+			Resource:  state.getPath() + "/" + item.Id.ValueString(),
 			Body:      "{}",
 		}
 	}